Fractions
First attempt at sharing contents on Github. This class can be used to
handle fractions (like you learned in elementary school). It calculates
the Greatest Common Divisor upon construction and upon changes to the
numerator and denominator, and provides the result of reducing the
fraction. Also methods are included for adding and subtracting
fractions.